// React/app/GerAdv_TS/Colaboradores/Interfaces/interface.Colaboradores.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IColaboradores {
// 202501251
    id: number;
 
	cargo: number,
	cliente: number,
	cidade: number,
	sexo: boolean,
	nome: string,
	cpf: string,
	rg: string,
	dtnasc: string,
	idade: number,
	endereco: string,
	bairro: string,
	cep: string,
	fone: string,
	observacao: string,
	email: string,
	cnh: string,
	class: string,
}

export interface IColaboradoresFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IColaboradoresIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}